# Optimize `array.filter` by attempting to no-op the result on consistent filters (up to 10% faster).