`array.filter` is a function that filters or retains elements from an `Array` based on some function.

We can speed it up when elements are consistently filtered:
- all `true`: We can return a shallow CoW copy. This is instantaneous, and will only be slow if it is modified after, forcing a full copy.
- all `false`: We can return an empty Array, which is instantaneous.
- The first X are `false`: We can make the resulting array smaller by X, speeding up allocation.

This PR was motivated by https://github.com/godotengine/godot/pull/102915 (https://github.com/godotengine/godot/pull/102915#issuecomment-2661478451).

## Benchmarks
I benchmarked a 10% improvement for affected cases in GDScript:
```gdscript
func _ready() -> void:
	const ITERATIONS = 100
	var array := range(100000)

	# Warmup
	var ignored = array.filter(func(v): v % 2)

	var begin := Time.get_ticks_usec()
	for i in ITERATIONS:
		var other = array.filter(func(v): v < 0)
	print("%dms for %d iterations (filter none)" % [(Time.get_ticks_usec() - begin) / 1000, ITERATIONS])

	begin = Time.get_ticks_usec()
	for i in ITERATIONS:
		var other = array.filter(func(v): v >= 0)
	print("%dms for %d iterations (filter all)" % [(Time.get_ticks_usec() - begin) / 1000, ITERATIONS])

	begin = Time.get_ticks_usec()
	for i in ITERATIONS:
		var other = array.filter(func(v): v < 500)
	print("%dms for %d iterations (filter first half)" % [(Time.get_ticks_usec() - begin) / 1000, ITERATIONS])

	begin = Time.get_ticks_usec()
	for i in ITERATIONS:
		var other = array.filter(func(v): v > 500)
	print("%dms for %d iterations (filter second half)" % [(Time.get_ticks_usec() - begin) / 1000, ITERATIONS])

	begin = Time.get_ticks_usec()
	for i in ITERATIONS:
		var other = array.filter(func(v): v % 2)
	print("%dms for %d iterations (filter interspersed)" % [(Time.get_ticks_usec() - begin) / 1000, ITERATIONS])
```

On `master`, this prints:
```
939ms for 100 iterations (filter none)
932ms for 100 iterations (filter all)
934ms for 100 iterations (filter first half)
932ms for 100 iterations (filter second half)
999ms for 100 iterations (filter interspersed)
```

On this PR, this prints:
```
908ms for 100 iterations (filter none)
904ms for 100 iterations (filter all)
903ms for 100 iterations (filter first half)
904ms for 100 iterations (filter second half)
967ms for 100 iterations (filter interspersed)
```

### Observations
The speedup is _much_ less than I expected. I attribute this to `callp` being slow, because GDScript establishes function frames, which appears to be the main bottleneck right now. Arguably, this makes this PR too complicated for little gain, and `GDScript` function calls need a look first. Still, 10% is nothing to sneeze at.

Interestingly, even the `interspersed` case is still sped up by `3.2%`, even though it's not an optimized case. I attribute this to re-using `result` and `ce` instead of re-assigning them each iteration.