@Ivorforce 2025-02-16
Converting to draft because I don't think it's worth merging this yet with such little gains.